+++ sdk/resource/os_test.go
@@ -20,7 +20,9 @@ import (
 
 	"github.com/stretchr/testify/require"
 
+	"go.opentelemetry.io/otel/attribute"
 	"go.opentelemetry.io/otel/sdk/resource"
+	semconv "go.opentelemetry.io/otel/semconv/v1.4.0"
 )
 
 func mockRuntimeProviders() {
@@ -83,4 +85,32 @@ func TestWithOS(t *testing.T) {
 		"os.type":        "linux",
 		"os.description": "Test",
 	}, toMap(res))
+}
+
+func TestMapRuntimeOSToSemconvOSType(t *testing.T) {
+	tt := []struct {
+		Name   string
+		Goos   string
+		OSType attribute.KeyValue
+	}{
+		{"Apple Darwin", "darwin", semconv.OSTypeDarwin},
+		{"DragonFly BSD", "dragonfly", semconv.OSTypeDragonflyBSD},
+		{"FreeBSD", "freebsd", semconv.OSTypeFreeBSD},
+		{"Linux", "linux", semconv.OSTypeLinux},
+		{"NetBSD", "netbsd", semconv.OSTypeNetBSD},
+		{"OpenBSD", "openbsd", semconv.OSTypeOpenBSD},
+		{"Oracle Solaris", "solaris", semconv.OSTypeSolaris},
+		{"Microsoft Windows", "windows", semconv.OSTypeWindows},
+		{"Unknown", "unknown", semconv.OSTypeKey.String("unknown")},
+		{"UNKNOWN", "UNKNOWN", semconv.OSTypeKey.String("unknown")},
+	}
+
+	for _, tc := range tt {
+		tc := tc
+
+		t.Run(tc.Name, func(t *testing.T) {
+			osTypeAttribute := resource.MapRuntimeOSToSemconvOSType(tc.Goos)
+			require.EqualValues(t, osTypeAttribute, tc.OSType)
+		})
+	}
 }